Handle null SubmittableId and code cleanning.
