+++ src/main/java/uk/ac/ebi/subs/progressmonitor/MonitorService.java
@@ -1,23 +1,89 @@
 package uk.ac.ebi.subs.progressmonitor;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 import uk.ac.ebi.subs.processing.ProcessingCertificate;
 import uk.ac.ebi.subs.processing.ProcessingCertificateEnvelope;
 import uk.ac.ebi.subs.processing.SubmissionEnvelope;
+import uk.ac.ebi.subs.repository.model.ProcessingStatus;
+import uk.ac.ebi.subs.repository.processing.SupportingSample;
+import uk.ac.ebi.subs.repository.processing.SupportingSampleRepository;
+import uk.ac.ebi.subs.repository.repos.status.ProcessingStatusRepository;
 
-/**
- * Created by davidr on 25/04/2017.
- */
-public interface MonitorService {
+import java.util.Date;
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Service
+public class MonitorService {
+    private static final Logger logger = LoggerFactory.getLogger(MonitorService.class);
+
+    @Autowired private SupportingSampleRepository supportingSampleRepository;
+    @Autowired private ProcessingStatusRepository processingStatusRepository;
+    @Autowired private SubmittablesBulkOperations submittablesBulkOperations;
 
     /**
      * store supporting information received from archives
      * @param submissionEnvelope
      */
-    void storeSupportingInformation(SubmissionEnvelope submissionEnvelope);
+    public void storeSupportingInformation(SubmissionEnvelope submissionEnvelope) {
+
+        final String submissionId = submissionEnvelope.getSubmission().getId();
+
+
+        List<SupportingSample> supportingSamples = submissionEnvelope.getSupportingSamples().stream()
+                .map(s -> new SupportingSample(submissionId, s))
+                .collect(Collectors.toList());
+
+        //store supporting info,
+        logger.info(
+                "storing supporting sample info for submission {}, {} samples",
+                submissionEnvelope.getSubmission().getId(),
+                supportingSamples.size()
+        );
+
+        supportingSampleRepository.save(supportingSamples);
+    }
 
     /**
      * update accessions + statuses using information in a processingCertificateEnvelop
      * @param processingCertificateEnvelope
      */
-    void updateSubmittablesFromCertificates(ProcessingCertificateEnvelope processingCertificateEnvelope);
+    public void updateSubmittablesFromCertificates(ProcessingCertificateEnvelope processingCertificateEnvelope) {
+
+        logger.info("received agent results for submission {} with {} certificates ",
+                processingCertificateEnvelope.getSubmissionId(), processingCertificateEnvelope.getProcessingCertificates().size());
+
+
+        for (ProcessingCertificate cert : processingCertificateEnvelope.getProcessingCertificates()) {
+
+            if (cert.getSubmittableId() == null) {
+                throw new NullSubmittableIdException("Processing Certificate Submittable Id can't be NULL.");
+            }
+
+            ProcessingStatus processingStatus = processingStatusRepository.findBySubmittableId(cert.getSubmittableId());
+
+            if (processingStatus == null) {
+                continue;
+            }
+
+            if (cert.getAccession() != null) {
+                processingStatus.setAccession(cert.getAccession());
+            }
+
+            processingStatus.setArchive(cert.getArchive().name());
+            processingStatus.setMessage(cert.getMessage());
+
+            processingStatus.setStatus(cert.getProcessingStatus());
+
+            processingStatus.setLastModifiedBy(cert.getArchive().name());
+            processingStatus.setLastModifiedDate(new Date());
+
+            processingStatusRepository.save(processingStatus);
+        }
+
+        submittablesBulkOperations.applyProcessingCertificates(processingCertificateEnvelope);
+    }
 }
