+++ src/main/java/uk/ac/ebi/subs/progressmonitor/NullSubmittableIdException.java
@@ -1,0 +1,8 @@
+package uk.ac.ebi.subs.progressmonitor;
+
+public class NullSubmittableIdException extends RuntimeException {
+
+    public NullSubmittableIdException(String message) {
+        super(message);
+    }
+}
