--- src/test/java/progressmonitor/ProgressMonitorTestConfiguration.java
@@ -1,9 +1,0 @@
-package progressmonitor;
-
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@EnableAutoConfiguration
-public class ProgressMonitorTestConfiguration {
-}
